chatsito-app/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.2.1
flask==3.1.2
flask-socketio==5.5.1
h11==0.16.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
werkzeug==3.1.3
wsproto==1.2.0
